+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## [5.48.2](https://github.com/plivo/plivo-java/tree/v5.48.2) (2026-06-10)
+**Feature - Profile API number of employees field support**
+- Added Number of Employees field support to Profile API
+
 ## [5.48.1](https://github.com/plivo/plivo-java/tree/v5.48.1) (2026-05-26)
 **Feature - Profile API DBA field support**
 - Added Doing Business As (DBA) field support to Profile API